Separeted each class in separeted files
