+++ lib/main.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'vban.dart';
+import 'editbuttonspage.dart';
 
 void main() {
   runApp(const MyApp());
@@ -127,192 +128,3 @@ class _MyHomePageState extends State<MyHomePage> {
   }
 }
 
-class EditButtonsPage extends StatefulWidget {
-  final List<String> name;
-  final List<String> cmd;
-  final Function(List<String>, List<String>) onSave;
-  final String ipAddress;
-  final int port;
-  final Function(String, int) onSaveSettings;
-
-  const EditButtonsPage({
-    super.key,
-    required this.name,
-    required this.cmd,
-    required this.onSave,
-    required this.ipAddress,
-    required this.port,
-    required this.onSaveSettings,
-  });
-
-  @override
-  _EditButtonsPageState createState() => _EditButtonsPageState();
-}
-
-class _EditButtonsPageState extends State<EditButtonsPage> {
-  late List<TextEditingController> nameControllers;
-  late List<TextEditingController> cmdControllers;
-
-  @override
-  void initState() {
-    super.initState();
-    nameControllers =
-        widget.name.map((name) => TextEditingController(text: name)).toList();
-    cmdControllers =
-        widget.cmd.map((cmd) => TextEditingController(text: cmd)).toList();
-  }
-
-  void _addButton() {
-    setState(() {
-      nameControllers.add(TextEditingController());
-      cmdControllers.add(TextEditingController());
-    });
-  }
-
-  void _removeButton(int index) {
-    setState(() {
-      nameControllers.removeAt(index);
-      cmdControllers.removeAt(index);
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Edit Buttons'),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.save),
-            onPressed: () {
-              List<String> newName =
-                  nameControllers.map((controller) => controller.text).toList();
-              List<String> newCmd =
-                  cmdControllers.map((controller) => controller.text).toList();
-              widget.onSave(newName, newCmd);
-              Navigator.pop(context);
-            },
-          ),
-          IconButton(
-            icon: const Icon(Icons.settings),
-            onPressed: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => SettingsPage(
-                    ipAddress: widget.ipAddress,
-                    port: widget.port,
-                    onSave: (newIp, newPort) {
-                      widget.onSaveSettings(newIp, newPort);
-                    },
-                  ),
-                ),
-              );
-            },
-          ),
-          IconButton(onPressed: _addButton, icon: const Icon(Icons.add)),
-        ],
-      ),
-      body: ListView.builder(
-        itemCount: nameControllers.length,
-        itemBuilder: (context, index) {
-          return Padding(
-            padding: const EdgeInsets.all(8.0),
-            child: Row(
-              children: [
-                Expanded(
-                  child: TextField(
-                    controller: nameControllers[index],
-                    decoration: const InputDecoration(labelText: 'Button Name'),
-                  ),
-                ),
-                const SizedBox(width: 8),
-                Expanded(
-                  child: TextField(
-                    controller: cmdControllers[index],
-                    decoration: const InputDecoration(labelText: 'Command'),
-                  ),
-                ),
-                IconButton(
-                  icon: const Icon(Icons.delete),
-                  onPressed: () => _removeButton(index),
-                ),
-              ],
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
-
-class SettingsPage extends StatefulWidget {
-  final String ipAddress;
-  final int port;
-  final Function(String, int) onSave;
-
-  const SettingsPage({
-    super.key,
-    required this.ipAddress,
-    required this.port,
-    required this.onSave,
-  });
-
-  @override
-  _SettingsPageState createState() => _SettingsPageState();
-}
-
-class _SettingsPageState extends State<SettingsPage> {
-  late TextEditingController ipController;
-  late TextEditingController portController;
-
-  @override
-  void initState() {
-    super.initState();
-    ipController = TextEditingController(text: widget.ipAddress);
-    portController = TextEditingController(text: widget.port.toString());
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: const Text('Paramètres'),
-        actions: [
-          IconButton(
-            icon: const Icon(Icons.save),
-            onPressed: () {
-              String newIp = ipController.text;
-              int newPort = int.tryParse(portController.text) ?? widget.port;
-
-              if (newIp.isNotEmpty && newPort > 0) {
-                widget.onSave(newIp, newPort);
-                Navigator.pop(context);
-              } else {
-                print("Invalid IP or Port");
-              }
-            },
-          ),
-        ],
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(16.0),
-        child: Column(
-          children: [
-            TextField(
-              controller: ipController,
-              decoration: const InputDecoration(labelText: 'Adresse IP'),
-              keyboardType: TextInputType.number,
-            ),
-            const SizedBox(height: 16),
-            TextField(
-              controller: portController,
-              decoration: const InputDecoration(labelText: 'Port'),
-              keyboardType: TextInputType.number,
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
